refactor: move initialization of storage instances
