+++ src/storage/index.js
@@ -1,0 +1,5 @@
+import { Queue } from '../Queue/queue.storage.js';
+import { Patients } from '../Resolution/resolution.storage.js';
+
+export const queue = new Queue();
+export const patients = new Patients();
